+++ specification/task/_types/TaskInfo.ts
@@ -17,11 +17,21 @@
  * under the License.
  */
 
-import { BaseNode } from '@spec_utils/BaseNode'
-import { Dictionary } from '@spec_utils/Dictionary'
-import { TaskId } from '@_types/common'
-import { TaskState } from './TaskState'
+import { Status } from '@task/_types/TaskStatus'
+import { HttpHeaders, Id } from '@_types/common'
+import { long } from '@_types/Numeric'
 
-export class TaskExecutingNode extends BaseNode {
-  tasks: Dictionary<TaskId, TaskState>
+export class Info {
+  action: string
+  cancellable: boolean
+  children?: Info[]
+  description?: string
+  headers: HttpHeaders
+  id: long
+  node: string
+  running_time_in_nanos: long
+  start_time_in_millis: long
+  status?: Status
+  type: string
+  parent_task_id?: Id
 }
