+++ specification/task/_types/TaskStatus.ts
@@ -17,9 +17,26 @@
  * under the License.
  */
 
-import { integer } from '@_types/Numeric'
+import { float, long } from '@_types/Numeric'
+import { Retries } from '@_types/Retries'
+import { Time } from '@_types/Time'
 
-export class TaskRetries {
-  bulk: integer
-  search: integer
+export class Status {
+  batches: long
+  canceled?: string
+  created: long
+  deleted: long
+  noops: long
+  failures?: string[]
+  requests_per_second: float
+  retries: Retries
+  throttled?: Time
+  throttled_millis: long
+  throttled_until?: Time
+  throttled_until_millis: long
+  timed_out?: boolean
+  took?: long
+  total: long
+  updated: long
+  version_conflicts: long
 }
